Merging Development (#10)
Sprint 1 and part of Sprint 2 development branch commits.

* Rename packages and moving classes.

For structure and organizational purposes, created new package for adapter patters, and renamed the packages to include group 2. Was done for both Test and Source packages.

* Added skeleton classes for the user package.

* Basic user classes finished.

Added an abstract user class ontop of the UserInterface. Finished JavaDocs on all finished methods and classes.

* Rename package to match other package formatting.

* Added user package UML diagram.

* Fix UML diagram aggrigation.

* Added abstract Log class.

* Changed classes to take Generic extended by LogAbstract rather than Log

* Added LogDecorator class.

* Added functioning decorator patterns to Log class.

* Renamed classes for better naming practice

* Updated logger package UML. Replaced pdf with png.